+++ app/src/main/java/labs/zero_one/roundo/MapKit.kt
@@ -1,9 +1,11 @@
 package labs.zero_one.roundo
 
 import android.content.Context
+import android.content.res.Resources
 import android.location.Location
 import android.os.Bundle
 import com.minemap.minemapsdk.MinemapAccountManager
+import com.minemap.minemapsdk.annotations.MarkerOptions
 import com.minemap.minemapsdk.camera.CameraPosition
 import com.minemap.minemapsdk.camera.CameraUpdateFactory
 import com.minemap.minemapsdk.geometry.LatLng
@@ -51,19 +53,26 @@ class MapKit(private val context: Context) {
         MinemapAccountManager.getInstance(
             context,
             context.getString(R.string.minemap_token),
-            "4807"
+            "4810"
         )
         mapView.onCreate(savedInstanceState)
         mapView.getMapAsync { newMap ->
             if (newMap == null) throw Exception(context.getString(R.string.err_map_init_failed))
             mineMap = newMap
             isMapInitialized = true
+            mineMap.setStyleUrl("http://minedata.cn/service/solu/style/id/4810")
             mineMap.uiSettings.isCompassEnabled = true
+            mineMap.uiSettings.setCompassMargins(
+                mineMap.uiSettings.compassMarginLeft,
+                (16 * Resources.getSystem().displayMetrics.density).toInt(),
+                (16 * Resources.getSystem().displayMetrics.density).toInt(),
+                mineMap.uiSettings.compassMarginBottom)
             mineMap.cameraPosition =
                 CameraPosition.Builder()
                     .target(LatLng(locationKit.lastLocation))
                     .zoom(16.0)
                     .build()
+
         }
     }
 
@@ -76,13 +85,37 @@ class MapKit(private val context: Context) {
      * @since 0.1.6
      */
     fun moveTo(location: Location){
-        if (isMapInitialized) {
-            mineMap.animateCamera(CameraUpdateFactory.newCameraPosition(
-                CameraPosition
-                    .Builder()
-                    .target(LatLng(location))
-                    .build()
-            ))
-        }
+        if (!isMapInitialized) return
+        mineMap.animateCamera(CameraUpdateFactory.newCameraPosition(
+            CameraPosition
+                .Builder()
+                .target(LatLng(location))
+                .build()
+        ))
+    }
+
+    /**
+     * 添加标记
+     *
+     * @param [marker] 标记
+     *
+     * @author lucka-me
+     * @since 0.1.7
+     */
+    fun add(marker: MarkerOptions) {
+        if (!isMapInitialized) return
+        mineMap.addMarker(marker)
+    }
+
+    /**
+     * 在指定位置添加标记
+     *
+     * @param [location] 目标位置
+     *
+     * @author lucka-me
+     * @since 0.1.7
+     */
+    fun addMarkAt(location: Location) {
+        add(MarkerOptions().position(LatLng(location)))
     }
 }